Renaming classes to controllers
