+++ public/index.php
@@ -2,8 +2,10 @@
 
 session_start();
 
+define('__BASE_DIR__', dirname(__DIR__) . '/');
+
 $config = require_once('../config.php');
 require_once '../vendor/autoload.php';
 
-$framework = new App\MasterController($config);
+$framework = new App\Controllers\MasterController($config);
 echo $framework->execute();
