+++ src/controllers/CommentController.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App;
+namespace App\Controllers;
 
 use PDO;
 
-class Comment {
+class CommentController {
     
     public function __construct($config) {
         $dbconfig = $config['database'];
