+++ src/controllers/MasterController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Controllers;
 
 class MasterController {
